+++ packages/react-ui/src/widgets/WrapperWidgetUI.js
@@ -110,9 +110,14 @@ const useStyles = makeStyles((theme) => ({
 
 function WrapperWidgetUI(props) {
   const wrapper = createRef();
+
   const [expandedInt, setExpandedInt] = useState(true);
-  const expanded = props.expandable !== false ? props.expanded ?? expandedInt : true;
-  const setExpanded = props.onExpendedChange ?? setExpandedInt;
+  const externalExpanded =
+    typeof props.expanded === 'boolean' && typeof props.onExpandedChange === 'function';
+  const expanded =
+    props.expandable !== false ? (externalExpanded ? props.expanded : expandedInt) : true;
+  const setExpanded = externalExpanded ? setExpandedInt : props.onExpandedChange;
+
   const [anchorEl, setAnchorEl] = useState(null);
   const classes = useStyles({ ...props, expanded });
   const open = Boolean(anchorEl);
@@ -246,6 +251,7 @@ function WrapperWidgetUI(props) {
 }
 
 WrapperWidgetUI.defaultProps = {
+  expanded: true,
   expandable: true,
   isLoading: false
 };
@@ -253,6 +259,8 @@ WrapperWidgetUI.defaultProps = {
 WrapperWidgetUI.propTypes = {
   title: PropTypes.string.isRequired,
   expandable: PropTypes.bool,
+  expanded: PropTypes.bool,
+  onExpandedChange: PropTypes.func,
   isLoading: PropTypes.bool,
   actions: PropTypes.arrayOf(
     PropTypes.shape({